fix(hot): Rename routes/Fetch/view.tsx to index.tsx
Login 라우터의 코드 베이스와 전혀 달랐네요 😅
